Add light radius to inspector UI

+++ src/entities/components/point_light.py
@@ -8,7 +8,7 @@ import numpy.typing as npt
 class PointLight:
     color: npt.NDArray[np.float32] = field(default_factory=lambda: np.array([1.0, 1.0, 1.0], dtype=np.float32))
     strength: np.float32 = np.float32(300.0)
-    radius: float = 0.05
+    radius: np.float32 = np.float32(0.05)
     enabled: bool = True
 
     shadow_map_fbo: int = 0
